modified to store doubles in .csv files.
Former-commit-id: 5ee7a80155371cc9fffd3d0d12b0ac4b3aada3ad

+++ STFC/HexitecGigEProject/trunk/HexitecGigE/mainwindow.h
@@ -58,7 +58,7 @@ private:
 //   void checkKeithleyConfiguration();
    void closeEvent(QCloseEvent *event);
    bool checkDAQChoice();
-   void writeCsv(QString filename, QVector<double> col0, int *col1, int numberOfBins);
+   void writeCsv(QString filename, QVector<double> col0, double *col1, int numberOfBins);
    QAction *startDAQAct;
    QAction *stopDAQAct;
